+++ lib/Table/TableStyle.ts
@@ -1,38 +1,28 @@
-import * as d3Scale from "d3-scale-chromatic";
 import groupBy from "lodash-es/groupBy";
 import { computed } from "mobx";
 import binarySearch from "terriajs-cesium/Source/Core/binarySearch";
-import Color from "terriajs-cesium/Source/Core/Color";
 import JulianDate from "terriajs-cesium/Source/Core/JulianDate";
 import TimeInterval from "terriajs-cesium/Source/Core/TimeInterval";
-import createColorForIdTransformer from "../Core/createColorForIdTransformer";
 import filterOutUndefined from "../Core/filterOutUndefined";
 import isDefined from "../Core/isDefined";
-import StandardCssColors from "../Core/StandardCssColors";
-import ColorMap from "../Map/ColorMap";
 import ConstantColorMap from "../Map/ConstantColorMap";
 import ConstantPointSizeMap from "../Map/ConstantPointSizeMap";
-import ContinuousColorMap from "../Map/ContinuousColorMap";
 import DiscreteColorMap from "../Map/DiscreteColorMap";
 import EnumColorMap from "../Map/EnumColorMap";
 import PointSizeMap from "../Map/PointSizeMap";
 import ScalePointSizeMap from "../Map/ScalePointSizeMap";
 import createCombinedModel from "../Models/createCombinedModel";
 import Model from "../Models/Model";
-import ModelPropertiesFromTraits from "../Models/ModelPropertiesFromTraits";
 import TableChartStyleTraits from "../Traits/TableChartStyleTraits";
-import TableColorStyleTraits, {
-  EnumColorTraits
-} from "../Traits/TableColorStyleTraits";
+import TableColorStyleTraits from "../Traits/TableColorStyleTraits";
 import TablePointSizeStyleTraits from "../Traits/TablePointSizeStyleTraits";
 import TableStyleTraits from "../Traits/TableStyleTraits";
 import TableTimeStyleTraits from "../Traits/TableTimeStyleTraits";
 import TableTraits from "../Traits/TableTraits";
+import TableColorMap from "./TableColorMap";
 import TableColumn from "./TableColumn";
 import TableColumnType from "./TableColumnType";
 
-const getColorForId = createColorForIdTransformer();
-const defaultColor = "yellow";
 const DEFAULT_FINAL_DURATION_SECONDS = 3600 * 24 - 1; // one day less a second, if there is only one date.
 
 interface TableModel extends Model<TableTraits> {
@@ -70,14 +60,19 @@ export default class TableStyle {
     );
   }
 
-  /** Hide style if number of colors (enumColors or numberOfBins) is less than 2. As a ColorMap with a single color isn't super useful. */
+  /** Hide style from "Display Variable" selector if number of colors (EnumColorMap or DiscreteColorMapw) is less than 2. As a ColorMap with a single color isn't super useful. */
   @computed
   get hidden() {
-    return (
-      this.styleTraits.hidden ??
-      ((this.isEnum && this.enumColors.length <= 1) ||
-        (!this.isEnum && this.numberOfBins <= 1))
-    );
+    if (isDefined(this.styleTraits.hidden)) return this.styleTraits.hidden;
+
+    if (this.colorMap instanceof ConstantColorMap) return true;
+
+    if (
+      (this.colorMap instanceof EnumColorMap ||
+        this.colorMap instanceof DiscreteColorMap) &&
+      this.colorMap.colors.length < 2
+    )
+      return true;
   }
 
   /**
@@ -264,16 +259,6 @@ export default class TableStyle {
     return this.xAxisColumn !== undefined && this.chartTraits.lines.length > 0;
   }
 
-  /** Column can use EnumColorMap if type is enum, region or text AND the number of unique values is less than (or equal to) the number of bins */
-  @computed get isEnum() {
-    return (
-      !!this.colorColumn &&
-      (this.colorColumn.type === TableColumnType.enum ||
-        this.colorColumn.type === TableColumnType.region) &&
-      this.colorColumn.uniqueValues.values.length <= this.binColors.length
-    );
-  }
-
   /** Style isSampled by default. TimeTraits.isSampled will be used if defined. If not, and color column is binary - isSampled will be false. */
   @computed get isSampled() {
     if (isDefined(this.timeTraits.isSampled)) return this.timeTraits.isSampled;
@@ -282,302 +267,16 @@ export default class TableStyle {
     return true;
   }
 
-  @computed
-  get colorPaletteName() {
-    const colorColumn = this.colorColumn;
-
-    if (colorColumn === undefined) {
-      return;
-    }
-
-    let paletteName = this.colorTraits.colorPalette;
-
-    if (
-      colorColumn.type === TableColumnType.enum ||
-      colorColumn.type === TableColumnType.region
-    ) {
-      // Enumerated values, so use a large, high contrast palette.
-      paletteName = paletteName || "HighContrast";
-    } else if (colorColumn.type === TableColumnType.scalar) {
-      if (paletteName === undefined) {
-        const valuesAsNumbers = colorColumn.valuesAsNumbers;
-        if (
-          valuesAsNumbers !== undefined &&
-          (valuesAsNumbers.minimum || 0.0) < 0.0 &&
-          (valuesAsNumbers.maximum || 0.0) > 0.0
-        ) {
-          // Values cross zero, so use a diverging palette
-          paletteName = "PuOr";
-        } else {
-          // Values do not cross zero so use a sequential palette.
-          paletteName = "Reds";
-        }
-      }
-    }
-
-    return paletteName;
-  }
-
-  @computed get fallbackColorPalette() {
-    return (
-      this.colorPaletteName
-        ?.split("-")
-        .map(color => Color.fromCssColorString(color) ?? Color.TRANSPARENT) ?? [
-        Color.TRANSPARENT
-      ]
+  @computed get tableColorMap() {
+    return new TableColorMap(
+      this.tableModel.name ?? this.tableModel.uniqueId,
+      this.colorColumn,
+      this.colorTraits
     );
   }
 
-  @computed
-  get numberOfBins(): number {
-    const colorColumn = this.colorColumn;
-    if (colorColumn === undefined) return this.binMaximums.length;
-    if (colorColumn.type === TableColumnType.scalar) {
-      return colorColumn.uniqueValues.values.length < this.binMaximums.length
-        ? colorColumn.uniqueValues.values.length
-        : this.binMaximums.length;
-    }
-    return this.binMaximums.length;
-  }
-
-  /**
-   * Gets the color to use for each bin. The length of the returned array
-   * will be equal to {@link #numberOfColorBins}.
-   */
-  @computed
-  get binColors(): readonly Readonly<Color>[] {
-    const numberOfBins = this.numberOfBins;
-
-    // Pick a color for every bin.
-    const binColors = this.colorTraits.binColors || [];
-
-    const colorScaleScheme = (d3Scale as any)[`scheme${this.colorPaletteName}`];
-
-    let colorScale = this.fallbackColorPalette;
-
-    if (Array.isArray(colorScaleScheme)) {
-      colorScale = colorScaleScheme[this.numberOfBins];
-      // If invalid numberOfBins - use largest number
-      if (!Array.isArray(colorScale)) {
-        colorScale = colorScaleScheme[colorScaleScheme.length - 1];
-      }
-    }
-
-    const result: Color[] = [];
-    for (let i = 0; i < numberOfBins; ++i) {
-      if (i < binColors.length) {
-        result.push(
-          Color.fromCssColorString(binColors[i]) ?? Color.TRANSPARENT
-        );
-      } else {
-        result.push(colorScale[i % colorScale.length]);
-      }
-    }
-    return result;
-  }
-
-  @computed
-  get binMaximums(): readonly number[] {
-    const colorColumn = this.colorColumn;
-    if (colorColumn === undefined) {
-      return this.colorTraits.binMaximums || [];
-    }
-
-    const binMaximums = this.colorTraits.binMaximums;
-    if (binMaximums !== undefined) {
-      if (
-        colorColumn.type === TableColumnType.scalar &&
-        colorColumn.valuesAsNumbers.maximum !== undefined &&
-        (binMaximums.length === 0 ||
-          colorColumn.valuesAsNumbers.maximum >
-            binMaximums[binMaximums.length - 1])
-      ) {
-        // Add an extra bin to accomodate the maximum value of the dataset.
-        return binMaximums.concat([colorColumn.valuesAsNumbers.maximum]);
-      }
-      return binMaximums;
-    } else if (this.colorTraits.numberOfBins === 0) {
-      return [];
-    } else {
-      // TODO: compute maximums according to ckmeans, quantile, etc.
-      const asNumbers = colorColumn.valuesAsNumbers;
-      const min = asNumbers.minimum;
-      const max = asNumbers.maximum;
-      if (min === undefined || max === undefined) {
-        return [];
-      }
-      const numberOfBins =
-        colorColumn.uniqueValues.values.length < this.colorTraits.numberOfBins
-          ? colorColumn.uniqueValues.values.length
-          : this.colorTraits.numberOfBins;
-      let next = min;
-      const step = (max - min) / numberOfBins;
-
-      const result: number[] = [];
-      for (let i = 0; i < numberOfBins - 1; ++i) {
-        next += step;
-        result.push(next);
-      }
-
-      result.push(max);
-
-      return result;
-    }
-  }
-
-  @computed
-  get enumColors(): readonly ModelPropertiesFromTraits<EnumColorTraits>[] {
-    if (this.colorTraits.enumColors.length > 0) {
-      return this.colorTraits.enumColors;
-    }
-
-    const colorColumn = this.colorColumn;
-    if (colorColumn === undefined) {
-      return [];
-    }
-
-    // Create a color for each unique value
-    const uniqueValues = colorColumn.uniqueValues.values;
-
-    let colorScale = StandardCssColors.highContrast;
-
-    if (isDefined(this.colorPaletteName)) {
-      colorScale =
-        (d3Scale as any)[`scheme${this.colorPaletteName}`] ??
-        this.fallbackColorPalette;
-    }
-
-    return uniqueValues.map((value, i) => {
-      return {
-        value: value,
-        color: colorScale[i % colorScale.length]
-      };
-    });
-  }
-
-  /**
-   * Gets an object used to map values in {@link #colorColumn} to colors
-   * for this style.
-   */
-  @computed
-  get colorMap(): ColorMap {
-    const colorColumn = this.colorColumn;
-    const colorTraits = this.colorTraits;
-
-    const nullColor = colorTraits.nullColor
-      ? Color.fromCssColorString(colorTraits.nullColor) ?? Color.TRANSPARENT
-      : Color.TRANSPARENT;
-
-    const minValue =
-      colorTraits.minimumValue ?? colorColumn?.valuesAsNumbers.minimum;
-    const maxValue =
-      colorTraits.maximumValue ?? colorColumn?.valuesAsNumbers.maximum;
-
-    if (colorColumn && colorColumn.type === TableColumnType.scalar) {
-      if (this.binMaximums.length > 0) {
-        return new DiscreteColorMap({
-          bins: this.binColors.map((color, i) => {
-            return {
-              color: color,
-              maximum: this.binMaximums[i],
-              includeMinimumInThisBin: false
-            };
-          }),
-          nullColor
-        });
-      } else if (
-        this.colorPaletteName &&
-        isDefined(minValue) &&
-        isDefined(maxValue)
-      ) {
-        const colorScale = (d3Scale as any)[
-          `interpolate${this.colorPaletteName}`
-        ];
-
-        if (typeof colorScale !== "function")
-          throw `Color palette "${this.colorPaletteName}" is not valid.`;
-
-        return new ContinuousColorMap({
-          colorScale,
-          minValue,
-          maxValue,
-          nullColor
-        });
-      }
-    } else if (colorColumn && this.isEnum) {
-      return new EnumColorMap({
-        enumColors: filterOutUndefined(
-          this.enumColors.map(e => {
-            if (e.value === undefined || e.color === undefined) {
-              return undefined;
-            }
-            return {
-              value: e.value,
-              color:
-                colorColumn.type !== TableColumnType.region
-                  ? Color.fromCssColorString(e.color) ?? Color.TRANSPARENT
-                  : this.regionColor
-            };
-          })
-        ),
-        nullColor: this.nullColor
-      });
-    } else {
-      // No column to color by, so use the same color for everything.
-      let color: Color | undefined;
-
-      const colorId = this.tableModel.uniqueId || this.tableModel.name;
-
-      // If colorColumn is of type region - use regionColor
-      if (colorColumn?.type === TableColumnType.region && this.regionColor) {
-        color = this.regionColor;
-      } else if (colorTraits.nullColor) {
-        color = Color.fromCssColorString(colorTraits.nullColor);
-      } else if (colorTraits.binColors && colorTraits.binColors.length > 0) {
-        color = Color.fromCssColorString(colorTraits.binColors[0]);
-      } else if (colorId) {
-        color = Color.fromCssColorString(getColorForId(colorId));
-      }
-
-      if (!color) {
-        color = Color.fromCssColorString(defaultColor);
-      }
-
-      // Use nullColor if colorColumn is of type `region`
-      let nullColor =
-        colorColumn?.type === TableColumnType.region
-          ? this.nullColor
-          : undefined;
-
-      return new ConstantColorMap({
-        color,
-        title: this.tableModel.name,
-        nullColor
-      });
-    }
-
-    // No column to color by, so use the same color for everything.
-    let color = Color.fromCssColorString(defaultColor);
-    const colorId = this.tableModel.uniqueId || this.tableModel.name;
-    if (colorTraits.nullColor) {
-      color = Color.fromCssColorString(colorTraits.nullColor);
-    } else if (this.binColors.length > 0) {
-      color = this.binColors[0];
-    } else if (colorId) {
-      color = Color.fromCssColorString(getColorForId(colorId));
-    }
-    return new ConstantColorMap({ color, title: this.tableModel.name });
-  }
-
-  @computed get nullColor() {
-    return this.colorTraits.nullColor
-      ? Color.fromCssColorString(this.colorTraits.nullColor) ??
-          Color.TRANSPARENT
-      : Color.TRANSPARENT;
-  }
-
-  @computed get regionColor() {
-    return Color.fromCssColorString(this.colorTraits.regionColor);
+  @computed get colorMap() {
+    return this.tableColorMap.colorMap;
   }
 
   @computed